Count number of correctly entered cards in Day 2 Exercise 11
I'd incorrectly placed an increment within a while loop, so that the number
of cards entered was incremented every time the suit was asked for, so was
being incremented even if an incorrect suit was entered. Corrected so that the
number of cards is only incremented after the suit has been entered correctly.