Alternative solution to Day 2 Exercise 11
My previous solution to Exercise 11 checked for a straight by looking for the
maximum and minimum differences between pairs of cards, and then separately
compared each card with all the other cards in the hand to check whether they
had the same rank, to look for pairs, three-of-a-kind, etc.

I now create a 13-element array containing the number of cards of each rank, and
use this array to check for straights and for repeated ranks. I think this is
clearer than the previous method, and it should also be more effecient, as I only
need to run through the hand once, and then run through the 13-element array
once (13 + 5 = 18 comparisons), rather than comparing each card with each other
card, and then checking through the resulting array (5^2 + 5 = 30) comparisons.
This new method also makes it easier to check for 10-J-Q-K-A straights.