Make DoublyLinkedList implement Iterator interface
I'd already added a next() method to DoublyLinkedList. This commit adds the relevant features to allow the DoublyLinkedList to be treated as a Java Iterator.

(It's probably not a good idea to have the class itself be an iterator - what if two bits of code want to iterate over the list at the same time - but it's simply for this illustration).